feat: add smart case matching
If the query contains all lowercase, then case is ignored. If the query
contains at least one uppercase letter, then case is considered during
the fuzzy matching.

The lowercase copy of the string is stored upon Candidate creation. It
is preferred to take twice the memory over calculating the lowercase
version on every filter iteration.