feat: update ranking algorithm
Inspired by selecta by Gary Bernhardt, the ranking algorithm has been
updated, with the special zf flavor added.

selecta uses a straightforward and efficient algorithm for ranking,
providing "bonuses" for patterns to rank results higher when those
patterns are used. zf extends this idea to include the following:

* The query string is tokenized on spaces. Each token is searched for
  inside the string. The sum of the rankings for each token is the
  candidate's score.
* The basename of the candidate is searched before the full path. If a
  match is found in the basename, then the candidate is given an
  additional bonus ranking.

The first (tokenization) makes it trivial to refine a search by
appending an additional query. The second (basenames) is optimal because
filenames in a project are typically unique.

This makes zf nearly ready for use! There are a few small tweaks and
optimizations remaining.