feat: add --lines option
Adds an -l --lines argument to specify the maximum number of lines to be
drawn, defaulting to 10. If the number of candidates is less than the
number of lines, then the number of lines is clamped to the number of
candidates.

To display full screen, set lines to a large number like 1000. The
number of lines will be clamped to the height of the terminal.

Note: zf does not listen for terminal resize events, so the height is
only recorded once.