feat: highlight all tokens
Improves highlighting by storing highlight ranges for all tokens. To not
put a fixed cap on the number of tokens and highlight ranges this does
require heap allocations.

Currently this leaks memory. On a project of about 12000 files memory
grew to about 200MB after a few minutes use. While zf is intended to
not run longer than 10 seconds on the regular run, this should still be
addressed in the future.