refactor: optimize highlight drawing code
This optimizes the drawing of highlights to no longer draw
character-by-character. A few bugs with highlight drawing are fixed,
but there are still a few edge cases to catch.

The refactor was an attempt to reduce flicker in the UI when drawing
highlights. The flicker still exists, but this seems to have helped
slightly.

Part of #6