feat(#3): prioritize on filename coverage
This further improves the ranking algorithm by prioritizing a match by
how much of the name it matched. As a concrete example, assume that
./nested/test.txt and ./my_test.txt exist. Previous to this commit, a
search for "test" would rank my_test.txt over test.txt because the rank
would be the same, so the shorter path would be shown.

With this change, test.txt would have a better rank because the query
"test" matches a greater portion of the filename.

This prioritization is only applied to filename matching for now, but
maybe it could be applied in general.

Closes #3