fix(#6): buffer output to reduce flicker
It turns out that the default TTY writer in Zig is not buffered, so each
write call required a syscall which led to the delay and visible flicker
in the UI.

This just wraps the writer in a BufferedWriter to buffer up to 4096
bytes before writing. Even with 50 lines of display and multiple
highlights, zf does not flicker anymore!

Closes #6