refactor: remove name from the Candidate struct
An attempt for optimization. If we can make the Candidate smaller we
might find better runtimes. Testing this commit with hyperfine shows
inconclusive results, but more work can be done after this to see if it
is a good route to take.

Even if it's not a speed up, this commit makes the code more readable so
it's a win