refactor: remove Ranges from candidate structs
Previously the calculation of the ranges were part of the fuzzy finding
algorithm. This is pointless, because you typically only need a range
for around 20 items (the number of displayed items is usually small,
while the full set of candidates could be in the 100,000s). This means
extra work and memory. Measured a small, yet noticeable difference after
removing Ranges from the struct.

At this point the library interface and the UI do not properly support
ranges.