refactor: lift tokenization to before ranking
Moves the tokenization step outside of the fuzzy algorithm. It now
expects the query to be tokenized. This allows the highlighting and
ranking algorithms to become separate and increase performance.

This does not have a measurable impact on performance, but it makes
the code easier to use as a library without requiring allocations.