refactor: rewrite terminal reading logic
Moves the terminal reading code to term.zig and rewrites to hopefully be
a bit more resilient against potentially bad input. Also returns slices
of input to better support reading codepoints and graphemes.

Currently it will either return a control code or a string, but not a
mixed buffer. I don't expect any mixed buffers to be needed, but this
can always be changed in the future.