feat(#14): add multiselect support
Adds the ability to select and deselect candidates. There are some rough
edges still, but for now this works well enough. Lines can be selected
down with Tab, and selected up with Shift Tab. Selecting a selected line
deselects it.

If any lines are selected, the number selected is displayed in the top
right of the UI.

Lines are written to stdout separated by newlines. This may be
configurable in the future.

Closes #14