feat: add pselect event loop
Rather than rely on the nodelay/blocking nature of a TTY read(), this
introduces an event loop based on pselect(2).

It currently blocks until there is available TTY input, or a SIGWINCH
signal is received.

There is still some room for improvements, but this provides a solid
base for extension like handling input on subprocesses for a preview
split.

Closes #40
Closes #39